Ismael Flores iaguilar@quarksoft.net * Se crea logica para endpoint de consulta de movimientos de estado cuenta con estructura de datos en BD actuales

+++ src/main/java/mx/com/nmp/pagos/mimonte/dto/conciliacion/MovimientosEstadoCuentaDTO.java
@@ -17,24 +17,24 @@ import java.util.List;
  */
 public class MovimientosEstadoCuentaDTO implements Comparable<MovimientosEstadoCuentaDTO> {
 
-	private Double total;
+	private Long total;
 	private List<MovimientoEstadoCuentaDTO> movimientos;
 
 	public MovimientosEstadoCuentaDTO() {
 		super();
 	}
 
-	public MovimientosEstadoCuentaDTO(Double total, List<MovimientoEstadoCuentaDTO> movimientos) {
+	public MovimientosEstadoCuentaDTO(Long total, List<MovimientoEstadoCuentaDTO> movimientos) {
 		super();
 		this.total = total;
 		this.movimientos = movimientos;
 	}
 
-	public Double getTotal() {
+	public Long getTotal() {
 		return total;
 	}
 
-	public void setTotal(Double total) {
+	public void setTotal(Long total) {
 		this.total = total;
 	}
